Move range validation to NumberControl
